+++ src/components/SearchInput/SearchInput.js
@@ -2,9 +2,18 @@ import { useEffect, useRef } from 'react';
 import { FaSearch } from 'react-icons/fa';
 import styles from './SearchInput.module.scss';
 
+/**
+ * Render the SearchInput component.
+ *
+ * @param {Props} props The props object.
+ * @param {string} props.value The search input value
+ * @param {(newValue: string) => void} props.onChange The search input onChange handler
+ * @returns {React.ReactElement} The SearchInput component.
+ */
 export default function SearchInput({ value, onChange }) {
   const input = useRef();
 
+  // Clear and focus the input on initial render
   useEffect(() => {
     input.current.value = '';
     input.current.focus();
