+++ src/components/SearchRecommendations/SearchRecommendations.module.scss
@@ -1,10 +1,11 @@
 .recommendations {
-    a {
-        color: var(--color-black);
+  a {
+    color: var(--color-black);
 
-        &:hover, &:focus {
-            color: var(--color-black);
-            text-decoration: underline;
-        }
+    &:hover,
+    &:focus {
+      color: var(--color-black);
+      text-decoration: underline;
     }
+  }
 }
