+++ src/components/SearchResults/SearchResults.module.scss
@@ -84,5 +84,5 @@
 }
 
 .no-results-text {
-    font-size: 3rem;
+  font-size: 3rem;
 }
